Déplacer code de téléchargement des fichiers espèces impactées dans un fichier dédié
